perf: Widen the scan key-range statistic to unsigned and partial reads
The integer value range folded out of parquet footers bounds a join key's
domain from above. It was only kept for signed columns read in full, which
left it unavailable for unsigned keys and for any scan whose footers were
sampled.

Unsigned columns are stored as `Int32`/`Int64` with unsigned-ordered bounds,
so their statistics are now reinterpreted through the matching unsigned type
and folded in `i128`. Signedness is read from the logical type, falling back
to the older converted type, which also closes a gap where a column annotated
only with the converted type was read as signed.

A sampled or partial read yields a range that can only be narrower than the
column's, which estimates the domain low and the join high, so it is kept
rather than discarded. `int_range_unknown` becomes `int_range_incomplete`,
which is what it has always meant.

+++ crates/polars-plan/src/plans/stats/mod.rs
@@ -100,19 +100,19 @@ pub struct ScanColumnStats {
     pub null_count: Card,
     /// Average width of one value in the source, in bytes.
     pub avg_byte_width: Option<f32>,
-    /// Inclusive value range of an integer column. Only set when every chunk of the
-    /// column was read and carried a min and a max, so the bound holds for the whole
-    /// column. Bounds the distinct count from above.
+    /// Inclusive value range of an integer column, folded over the chunks that were
+    /// read. Bounds the distinct count from above when those chunks cover the column,
+    /// and understates it otherwise, since a chunk that was skipped can only widen
+    /// the range.
     #[cfg_attr(feature = "serde", serde(default))]
-    pub int_range: Option<(i64, i64)>,
+    pub int_range: Option<(i128, i128)>,
 }
 
 impl ScanColumnStats {
-    /// Values the column could hold, from its integer range. An upper bound on the
-    /// distinct count, saturating rather than overflowing on a full-width range.
+    /// Values the column could hold, from its integer range.
     pub fn int_domain(&self) -> Option<f64> {
         let (min, max) = self.int_range?;
-        (max >= min).then(|| (max as i128 - min as i128 + 1) as f64)
+        (max >= min).then(|| (max - min + 1) as f64)
     }
 }
